All views passed once.

+++ src/django/VLE/views/role.py
@@ -85,7 +85,7 @@ class RoleView(viewsets.ViewSet):
                 return response.not_found('Course does not exist.')
 
             perms = permissions.get_permissions(user, int(course_id))
-            if not perms:
+            if perms is None:
                 return response.forbidden('You are not participating in this course')
 
             return response.success({'role': perms})
@@ -236,6 +236,9 @@ class RoleView(viewsets.ViewSet):
             return response.forbidden(description="You have no access to this course")
         elif not role.can_edit_course_roles:
             return response.forbidden(description="You have no permissions to delete this course role.")
+
+        if name in ['Student', 'TA', 'Teacher']:
+            return response.bad_request("You cannot delete the base roles of student, TA and teacher.")
 
         Role.objects.get(name=name, course=pk).delete()
         return response.success(description='Succesfully deleted role from course.')
